+++ min/index.js
@@ -1,6 +1,0 @@
-module.exports =  {
-  html: require('./html'),
-  css:  require('./css'),
-  js:   require('./js'),
-  img:  require('./img')
-};