+++ utils/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-  resolve:     require('./resolve'),
-  log:         require('./log'),
-  interpolate: require('./interpolate')
-};